requirements (2).txt
Flask>=2.2.2
torch===1.7.1
transformers~=4.2.0
gunicorn>=20.0.0
six==1.16.0
jinja2<3.1.0
google-cloud-storage
pandas
openpyxl
yake